# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SRB Motor API",
    "description": "Dokumentasi API lengkap untuk aplikasi SRB Motor (Mobile & Integrasi). Menggunakan Laravel Sanctum untuk keamanan rute terproteksi.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Local Development Server (artisan serve)"
    },
    {
      "url": "http://srbmotor.test/api",
      "description": "Local Laragon Host"
    }
  ],
  "paths": {
    "/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Registrasi Pengguna Baru",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "email": {
                    "type": "string",
                    "example": "john@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  },
                  "phone": {
                    "type": "string",
                    "example": "081234567890"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registrasi Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Registrasi berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    },
                    "access_token": {
                      "type": "string",
                      "example": "1|abcdefg..."
                    },
                    "token_type": {
                      "type": "string",
                      "example": "Bearer"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validasi Gagal (e.g. Email sudah terdaftar, password kurang dari 8 karakter)"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login Tradisional (Email & Password)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "john@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Login berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    },
                    "access_token": {
                      "type": "string",
                      "example": "2|hijklmn..."
                    },
                    "token_type": {
                      "type": "string",
                      "example": "Bearer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Password Salah"
          },
          "404": {
            "description": "Email Tidak Terdaftar"
          }
        }
      }
    },
...
    "/profile": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Perbarui Profil Pengguna",
        "description": "Gunakan Content-Type multipart/form-data untuk menyertakan file foto profil.",
        "security": [
          {
            "sanctum": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Doe Updated"
                  },
                  "phone": {
                    "type": "string",
                    "example": "081299998888"
                  },
                  "nik": {
                    "type": "string",
                    "example": "3201234567890002"
                  },
                  "alamat": {
                    "type": "string",
                    "example": "Jl. Keadilan Baru No. 10"
                  },
                  "occupation": {
                    "type": "string",
                    "example": "Wirausaha"
                  },
                  "profile_photo": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profil Berhasil Diperbarui",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Profil berhasil diperbarui"
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validasi Gagal"
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
...
    "/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logout / Hapus Token Aktif",
        "security": [
          {
            "sanctum": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Logout berhasil"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
    "/motors": {
      "get": {
        "tags": [
          "Motor Catalog"
        ],
        "summary": "Dapatkan Daftar Semua Motor",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Cari nama motor",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "brand",
            "in": "query",
            "description": "Filter merek motor (contoh: Honda, Yamaha)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter tipe motor (contoh: Matic, Manual)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sukses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Motor"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/motors/{id}": {
      "get": {
        "tags": [
          "Motor Catalog"
        ],
        "summary": "Detail Motor Spesifik",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID Motor",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sukses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Motor"
                }
              }
            }
          },
          "404": {
            "description": "Motor Tidak Ditemukan"
          }
        }
      }
    },
...
    "/branches/options": {
      "get": {
        "tags": [
          "Branches"
        ],
        "summary": "Dapatkan Opsi Nama & Kode Cabang (Untuk Dropdown Form)",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Kirim 'service' untuk memfilter cabang yang menyediakan servis",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sukses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "options": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      },
                      "example": {
                        "SSM_BDG": "SRB Motor Bandung",
                        "SSM_JKT": "SRB Motor Jakarta"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "phone": {
            "type": "string",
            "example": "081234567890"
          },
          "nik": {
            "type": "string",
            "example": "3201234567890001"
          },
          "alamat": {
            "type": "string",
            "example": "Jl. Raya Motor No. 123"
          },
          "occupation": {
            "type": "string",
            "example": "Karyawan Swasta"
          },
          "role": {
            "type": "string",
            "example": "user"
          },
          "profile_photo_path": {
            "type": "string",
            "example": "http://localhost:8000/storage/profile-photos/123.jpg"
          }
        }
      },
...
    },
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "description": "Masukkan Token Sanctum Bearer di sini."
      }
    }
  }
}